+++ src/main/java/svenhjol/charm/module/biome_dungeons/DungeonGenerator.java
@@ -1,13 +1,13 @@
 package svenhjol.charm.module.biome_dungeons;
 
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.levelgen.feature.structures.StructureTemplatePool;
 import svenhjol.charm.Charm;
 import svenhjol.charm.world.CharmGenerator;
 import svenhjol.charm.world.CharmStructure;
 
 import java.util.ArrayList;
 import java.util.List;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.level.levelgen.feature.structures.StructureTemplatePool;
 
 public class DungeonGenerator extends CharmGenerator {
     public static StructureTemplatePool BADLANDS_POOL;
@@ -20,6 +20,7 @@ public class DungeonGenerator extends CharmGenerator {
     public static StructureTemplatePool SAVANNA_POOL;
     public static StructureTemplatePool SNOWY_POOL;
     public static StructureTemplatePool TAIGA_POOL;
+    public static StructureTemplatePool EMPTY_POOL;
 
     public static List<CharmStructure> BADLANDS_DUNGEONS = new ArrayList<>();
     public static List<CharmStructure> DESERT_DUNGEONS = new ArrayList<>();
@@ -43,5 +44,6 @@ public class DungeonGenerator extends CharmGenerator {
         SAVANNA_POOL = registerPool(new ResourceLocation(Charm.MOD_ID, "dungeons/savanna/starts"), SAVANNA_DUNGEONS);
         SNOWY_POOL = registerPool(new ResourceLocation(Charm.MOD_ID, "dungeons/snowy/starts"), SNOWY_DUNGEONS);
         TAIGA_POOL = registerPool(new ResourceLocation(Charm.MOD_ID, "dungeons/taiga/starts"), TAIGA_DUNGEONS);
+        EMPTY_POOL = emptyPool(new ResourceLocation(Charm.MOD_ID, "dungeons/empty/starts"));
     }
 }